+++ web/src/main/java/com/web/Constants.java
@@ -1,0 +1,16 @@
+package com.web;
+
+public class Constants {
+    public final static String ENFORCER_CONFIG = "enforcer_config";
+    public final static String PX_USERNAME = "pxUser";
+    public final static String PX_PASSWORD = "1234";
+    public final static String JSON_SUFFIX = ".json";
+    public final static String PX_PREFIX = "PX";
+    public final static String THIRD_PARTY_SENSOR_URL_TEMPLATE = "//client.px-cloud.net/%s/main.min.js";
+    public final static String FIRST_PARTY_SENSOR_SUFFIX = "/init.js";
+    public final static String APP_ID_KEY = "app_id";
+    public final static String PX_APP_ID_FIELD = "px_app_id";
+    public final static String SENSOR_SRC_KEY = "sensor_src_url";
+    public final static String RESOURCES_RELATIVE_PATH = "src/main/resources/";
+    public final static String COMPILED_FILES_BASIC_PATH = "/target/classes";
+}