+++ web/src/main/java/com/web/PXFilter.java
@@ -1,0 +1,53 @@
+package com.web;
+
+import com.perimeterx.api.PerimeterX;
+import com.perimeterx.models.PXContext;
+import com.perimeterx.models.exceptions.PXException;
+
+import javax.servlet.*;
+import javax.servlet.annotation.WebFilter;
+import javax.servlet.http.HttpServletRequest;
+import javax.servlet.http.HttpServletResponse;
+import javax.servlet.http.HttpServletResponseWrapper;
+import java.io.IOException;
+
+import static com.web.Utils.setDefaultPageAttributes;
+
+@WebFilter("/*")
+public class PXFilter implements Filter {
+    private PerimeterX pxFilter;
+    private Config config;
+
+    public void init(FilterConfig filterConfig) {
+        try {
+            config = new Config();
+            pxFilter = new PerimeterX(config.getPxConfiguration());
+
+        } catch (PXException e) {
+            e.printStackTrace();
+            System.exit(1);
+        }
+    }
+
+    public void doFilter(ServletRequest request, ServletResponse response, FilterChain filterChain) throws IOException, ServletException {
+        try {
+
+            PXContext context = pxFilter.pxVerify((HttpServletRequest) request, new HttpServletResponseWrapper((HttpServletResponse) response));
+            setDefaultPageAttributes((HttpServletRequest) request, config);
+
+            if (context != null && context.isRequestLowScore()) {
+                filterChain.doFilter(request, response);
+            }
+
+
+        } catch (PXException e) {
+            filterChain.doFilter(request, response);
+        }
+
+    }
+
+    @Override
+    public void destroy() {
+
+    }
+}